new-mcp-client/requirements.txt
streamlit>=1.30.0
anthropic>=0.49.0
aiohttp>=3.8.4
python-dotenv>=1.0.0 